abstract away WCsettings
